broke out animation and postioned pieces
