+++ src/ui/registration/components/entropy.tsx
@@ -1,11 +1,11 @@
-import React, {useEffect, useRef} from 'react'
-import { StyleSheet, View, Text, Animated } from 'react-native'
+import React from 'react'
+import { StyleSheet, View, Text } from 'react-native'
 import { Container } from 'src/ui/structure/'
 import { MaskedImageComponent } from 'src/ui/registration/components/maskedImage'
 import I18n from 'src/locales/i18n'
 import strings from '../../../locales/strings'
 import { Typography, Colors } from 'src/styles'
-import { HandIcon, SplashIcon } from 'src/resources/index'
+import { HandAnimationComponent } from './handAnimation'
 
 interface Props {
   addPoint: (x: number, y: number) => void
@@ -29,37 +29,9 @@ const styles = StyleSheet.create({
   },
 })
 
-const usePulse = (pulseTiming: (val: Animated.Value) => Animated.CompositeAnimation) => {
-  const pulseValue = useRef(new Animated.Value(0)).current
-
-  const pulse = () => pulseTiming(pulseValue).start(pulse)
-
-  useEffect(() => {
-    const timeout = setTimeout(pulse, 0)
-    return () => clearTimeout(timeout)
-  })
-
-  return pulseValue
-}
-
-const handTiming = (val: Animated.Value) =>
-  Animated.sequence([
-    Animated.timing(val, { toValue: 1 }),
-    Animated.timing(val, { toValue: 0 })
-  ])
-
-const splashTiming = (val: Animated.Value) =>
-  Animated.sequence([
-    Animated.timing(val, {toValue: 1}),
-    Animated.timing(val, {toValue: 0})
-  ])
-
 export const EntropyComponent: React.SFC<Props> = props => {
   const { progress, addPoint } = props
   
-  const handOpacity = usePulse(handTiming)
-  const splashOpacity = usePulse(splashTiming)
-
   const msg =
     progress === 0
     ? I18n.t(strings.FOR_SECURITY_PURPOSES_WE_NEED_SOME_RANDOMNESS) +
@@ -75,14 +47,7 @@ export const EntropyComponent: React.SFC<Props> = props => {
       <View style={{ width: '100%' }}>
         {
           progress === 0
-          ? <View>
-              <Animated.View style={{ splashOpacity }}>
-                <SplashIcon />
-              </Animated.View>
-              <Animated.View style={{ handOpacity }}>
-                <HandIcon />
-              </Animated.View>
-            </View>
+          ? <HandAnimationComponent />
           : <MaskedImageComponent disabled={progress === 1} addPoint={addPoint} />
         }
       </View>
